refactor(ui): extract presentation_list drag helpers to sibling module
After Task 3 added handle_search_drop_at_fixed + render_list_spacer,
presentation_list.rs grew to 1029 lines, over the project's 1000-line
hard cap (CLAUDE.md). Move the drag-drop helpers and entry-payload
helpers into a new private sibling module presentation_list_drag.rs,
mirroring the slide_list_utils.rs pattern. presentation_list.rs is
now ~754 lines. No behavior change.
